Minor Bug Fixes and Design Changes

+++ Blogwrite/src/components/post-form/PostForm.jsx
@@ -8,6 +8,20 @@ import { useSelector } from "react-redux";
 
 export default function PostForm({post}) {
 
+    const slugTransform = useCallback((value) => {
+        if (value && typeof value === "string"){
+            return value
+                .trim()
+                .toLowerCase()
+                .replace(/[^a-zA-Z\d\s]+/g, "-")
+                .replace(/\s/g, "-");
+
+                console.log("inside slugTransform")
+            }
+    
+        return "";
+    }, []);
+
 
     // useForm hook to initialize form's feilds
 
@@ -17,10 +31,13 @@ export default function PostForm({post}) {
             defaultValues:{ 
                 //either new or from post object
                 title: post?.title || "",
-                slug: post?.slug || "",
+                slug: post?.$id ? slugTransform(post.title) : "",
                 content: post?.content || '',
+                status: post?.status || "active",
             },
-         })
+         });
+
+
     const navigate=useNavigate()
     
     const userData=useSelector((state)=> {
@@ -28,33 +45,41 @@ export default function PostForm({post}) {
         return state.auth.userData
     })
 
-    const submit = async(data)=>{
+    const submit = async (data)=>{
 
-        console.log('User Data:', userData);
+        console.log('User Data inside submit:', userData);
+        /*
+
         if (!userData || !userData.$id) {
             console.error('User ID is missing!');
             return; // Exit the function if userId is not available
         }
+        */
 
         // Check if post exists: EDIT MODE
         if (post){
+            console.log('In edit mode')
 
             //1. Image handling
             const file = data.image[0] ? await appwriteService.uploadFile(data.image[0]):null
 
             //if new image is uploaded---> removing old image 
-            if (file){
+            if (file  && post.featuredImage){
                 appwriteService.deleteFile(post.featuredImage);
             }
 
             //2.  Post update
             // Form data is sent to appwrite to update existing post, waits 
-            const dbPost = await appwriteService.updatePost(post.$id, {
-                ...data,
-                // if new image was uploaded, add to new post data
-                featuredImage: file ? file.$id : undefined,
+            let dbPost;
+            try {
+                dbPost = await appwriteService.updatePost(post.$id, {
+                    ...data,
+                    featuredImage: file ? file.$id : undefined,
+                });
+            } catch (error) {
+                console.error("Error updating post:", error);
+            }
 
-            });
 
             if (dbPost){
                 navigate(`/post/${dbPost.$id}`);
@@ -81,19 +106,7 @@ export default function PostForm({post}) {
         }              
     };
 
-        const slugTransform = useCallback((value) => {
-            if (value && typeof value === "string"){
-                return value
-                    .trim()
-                    .toLowerCase()
-                    .replace(/[^a-zA-Z\d\s]+/g, "-")
-                    .replace(/\s/g, "-");
-
-                    console.log("inside slugTransform")
-                }
         
-            return "";
-        }, []);
 
             //When you type something into the Title input,
             // the form is "watching" this field for any changes
@@ -121,6 +134,7 @@ export default function PostForm({post}) {
                     label="Title :"
                     placeholder="Title"
                     className="mb-4"
+                    labelcss="text-[#EEF5DB]"
                     {...register("title", { required: true })}
                 />
 
@@ -130,6 +144,7 @@ export default function PostForm({post}) {
                     label="Slug :"
                     placeholder="Slug"
                     className="mb-4"
+                    labelcss="text-[#EEF5DB]"
                     {...register("slug", { required: true })}
 
                     // on Input is for manual updating set values
@@ -138,7 +153,7 @@ export default function PostForm({post}) {
                     }}
                 />
 
-                <RTE label="Content :" name="content" control={control} defaultValue={getValues("content")} />
+                <RTE label="Content :" name="content" control={control} defaultValue={getValues("content")} labelcss="text-[#EEF5DB]" />
             </div>
 
             <div className="w-1/3 px-2">
@@ -146,6 +161,7 @@ export default function PostForm({post}) {
                     label="Featured Image :"
                     type="file"
                     className="mb-4"
+                    labelcss="text-[#EEF5DB]"
                     accept="image/png, image/jpg, image/jpeg, image/gif"
                     {...register("image", { required: !post })}
                 />
@@ -167,7 +183,7 @@ export default function PostForm({post}) {
                     {...register("status", { required: true })}
                 />
 
-                <Button type="submit" bgColor={post ? "bg-green-500" : undefined} className="w-full">
+                <Button onClick={()=>{console.log('btn clicked')}} type="submit" bgColor={post ? "bg-green-500" : undefined } className="w-full text-xl">
                     {post ? "Update" : "Submit"}
                 </Button>
             </div>
